+++ platform/frontend/src/app/settings/identity-providers/_parts/create-identity-provider-dialog.ee.tsx
@@ -68,6 +68,7 @@ export function CreateIdentityProviderDialog({
               oidcConfig: {
                 issuer: "",
                 pkce: true,
+                enableRpInitiatedLogout: true,
                 clientId: "",
                 clientSecret: "",
                 discoveryEndpoint: "",
@@ -77,6 +78,7 @@ export function CreateIdentityProviderDialog({
                   email: "email",
                   name: "name",
                 },
+                overrideUserInfo: true,
               },
             }),
       }),