+++ platform/frontend/src/app/settings/identity-providers/_parts/edit-identity-provider-dialog.ee.tsx
@@ -53,6 +53,7 @@ export function EditIdentityProviderDialog({
       oidcConfig: {
         issuer: "",
         pkce: true,
+        enableRpInitiatedLogout: true,
         clientId: "",
         clientSecret: "",
         discoveryEndpoint: "",
@@ -107,9 +108,10 @@ export function EditIdentityProviderDialog({
               },
             }
           : {
-              oidcConfig: provider.oidcConfig || {
+              oidcConfig: {
                 issuer: "",
                 pkce: true,
+                enableRpInitiatedLogout: true,
                 clientId: "",
                 clientSecret: "",
                 discoveryEndpoint: "",
@@ -120,6 +122,7 @@ export function EditIdentityProviderDialog({
                   name: "name",
                 },
                 overrideUserInfo: true,
+                ...provider.oidcConfig,
               },
             }),
       });